requirements.txt
Flask
Flask-CORS
python-dotenv
openai
pytesseract
Pillow
reportlab
pandas
openpyxl